requirements.txt
# Core dependencies
numpy==1.23.5
pandas>=1.5.0,<3.0.0
scipy>=1.6.0,<2.0.0

# Machine learning
scikit-learn==1.1.2
lightgbm==4.6.0
joblib==1.5.1
tensorflow-intel==2.17.1

# ONNX conversion
onnx>=1.14.1
onnxmltools>=1.8.1